Code/RestrictionScraper/requirements.txt
pytesseract==0.3.8
elasticsearch==7.14.0
pandas==1.3.2
pdfminer.six==20201018
datefinder==0.7.1
requests==2.26.0
pywebhdfs==0.4.1
beautifulsoup4==4.9.3
Image==1.5.33
pdfminer==20191125
Pillow==8.3.1
